[02-RUST.RS] .

+++ projects/comparar/find-factorial-of-a-number/02-rust.rs
@@ -9,7 +9,7 @@ fn main() {
   println!("Factorial({}) = {}", number, factorial(number));
 }
 
-fn factorial(n :i32) -> i32 {
+fn factorial(n :u64) -> u64 {
   if n == 1 { return 1; }
 
   return n * factorial(n-1);
